pyproject.toml
[project]
name = "image-processor"
version = "0.1.0"
description = "Django image processing application with timeline visualization"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "django-cors-headers>=4.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
